Add a note about asynchronous action queueing to readme
I'm bailing on asynchronous _processing_, since it would require synchronization over both the state and all history/gesturing vars,
and at that point, the only async processing available is `State::Update`, which is generally so cheap that the synchonization overhead
would likely be greater, as well as introducing complexity.

If I ever implement async action processing, it has to be done in a way that maintains the invariant that the store/gesture/history does
_not_ change within any single frame.
So likely a mutex involved.