Default-initialize store history with the main application store internally, rather than requiring the store be passed in.
History is already implicitly tied to the main `AppStore` instance in its `FinalizeGesture` method,
and it's currently only used with `AppStore`, so the old initializer was false advertisement of decoupling frrom a specific store.

Also, this is another callsite that required explicit knowledge of `AppStore` from the caller.
Better to just tie it to the store internally, and callers can simply reinstantiate with the default ctor to clear history.