Instead `Store` holding a single `immer::map` of `PrimitiveVariant`s, it now holds separate maps for each primitive type.
This has a lower memory footprint, since a variant instance is big enough to hold its largest element.
In this case, this meant e.g. a `bool` store value was the size of a `string`.

This also avoids wrapping and unwrapping primitive types into variants for every set/get.

The downside is running `immer::diff` over five maps instead of one for each patch.