Tranition all `PrimitiveVector`s to the new `immer::vector`-backed `PrimitiveVec`.
This turned into quite a long thread to pull, and includes lots of changes along with it.
Most notable, `PatchOp` received a lot of changes, with new vector/set-specific `PatchOpType`s,
and the `PatchOps` map type now holds a _vector_ of `PatchOp`s instead of just one,
since a patch op is no longer assumed to act on a single primitive type, so it can e.g. push
multiple items to a single container in a single patch.

Paths in patches now always point to a container, _except_ for vector set, which still appends
the index to set.