Remove `static` from  inline defs _not_ in classes to signal they can be used across translation units.
Counterintuitively (to me), this actually resulted in a somewhat _larger_ binary size.
However, this must be due to the compiler now inlining more calls than it was allowed to before,
and thus it decided the extra bloat is worth the perf savings from inlining, so this should
generally be a good thing!